split playback control sub-components by input data fetching approach
