+++ frontend/media-playback/controls/sub-components/using-props/BasicControls.tsx
