+++ frontend/media-playback/controls/sub-components/using-props/PlaybackRateSlider.tsx
