+++ frontend/media-playback/controls/sub-components/using-props/WaveFormView/SimplePointMarker.ts
