+++ frontend/media-playback/controls/sub-components/using-props/WaveFormView/WaveformView.tsx
