+++ frontend/media-playback/controls/sub-components/using-props/WaveformViewZoomControls.tsx
