+++ frontend/media-playback/controls/sub-components/using-store/PlaybackProgressBar.tsx
